confirm delete modal implemented

+++ src/pages/Dashboard.js
@@ -1,7 +1,7 @@
 import { auth, db } from "../utils/firebase";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { useNavigate, useLocation } from "react-router-dom";
-import { useEffect, useState } from "react";
+import { useEffect, useState, Fragment } from "react";
 import {
   collection,
   deleteDoc,
@@ -14,9 +14,12 @@ import {
 } from "firebase/firestore";
 import UserPost from "../components/UserPost";
 import { BsTrash3 } from "react-icons/bs";
-import { AiOutlineCloseSquare } from "react-icons/ai";
 import { CiEdit } from "react-icons/ci";
 import { Link } from "react-router-dom";
+import DeleteModal from "../components/DeleteModal";
+// React Toastify
+import { toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 //=====================================
 // to sign user out, all you have to do is auth.signOut() which is provided through firebase
 
@@ -25,6 +28,7 @@ function Dashboard() {
   const [user, loading] = useAuthState(auth);
   const [userPosts, setUserPosts] = useState([]);
   const [showModal, setShowModal] = useState(false);
+  const [selectedPostId, setSelectedPostId] = useState(null);
 
   // GET USERS POSTS
   const getUserPosts = async () => {
@@ -38,11 +42,28 @@ function Dashboard() {
     return posts;
   };
 
-  // DELETE A POST
-  const deletePost = async (id) => {
-    const docRef = doc(db, "posts", id);
+
+  const deletePost = async () => {
+    if (!selectedPostId) return;
+    const docRef = doc(db, "posts", selectedPostId);
     await deleteDoc(docRef);
+    setSelectedPostId(null);
+    closeModal();
+    toast.success("Post Deleted", {
+      position: toast.POSITION.BOTTOM_CENTER,
+      autoClose: 3000,
+    });
   };
+
+  // MODAL
+
+  function closeModal() {
+    setShowModal(false);
+  }
+
+  function openModal() {
+    setShowModal(true);
+  }
 
   //=============================
 
@@ -57,7 +78,7 @@ function Dashboard() {
   }, [user, loading]);
 
   return (
-    <main>
+    <main className="min-h-screen">
       <div className="h-40 my-4 flex justify-center items-center bg-gray-300">
         <h1 className="text-bold text-2xl text-zinc-600">Your Posts</h1>
       </div>
@@ -67,15 +88,22 @@ function Dashboard() {
           <UserPost post={post} key={post.id}>
             <div className="flex gap-2 pt-4  mt-4">
               <button
-                onClick={() => deletePost(post.id)}
+                onClick={() => {
+                  setSelectedPostId(post.id);
+                  openModal();
+                }}
                 className="font-bold bg-transparent text-zinc-600 shadow-md bg-zinc-100  py-2 px-6 rounded-xl text-sm flex items-center justify-center gap-2
-                transition transform hover:scale-105"
+              transition transform hover:scale-105"
               >
                 <BsTrash3 className="text-2xl" />
                 Delete
               </button>
 
-
+              <DeleteModal
+                showModal={showModal}
+                closeModal={closeModal}
+                deletePost={deletePost}
+              />
 
               <Link to={`/editpost/${post.id}`}>
                 <button
@@ -90,8 +118,6 @@ function Dashboard() {
           </UserPost>
         ))}
       </div>
-
-     
 
       <div className="flex justify-center items-center pb-8 pt-8">
         <button
